Simplify
Remove backend module and incorporate fetching code etc into MainView.js for simplicity
