+++ EVOLVE UI Kit.unoproj
@@ -6,7 +6,6 @@
   ],
   "Includes": [
     "*",
-    "Modules/*.js:Bundle",
     "Assets/Shoe Images/*:Bundle"
   ]
 }
